go_busybox_command -> go_busybox_library
bazel has migrated to using go_library even for packages with a main,
and generating a go_binary that points to the library. Our rule needs
access to the srcs and deps, so call it go_busybox_library and generate
a go_library standalone rule rather than a go_binary one.

This will also keep gazelle happy.

Signed-off-by: Chris Koch <chrisko@google.com>