bazel: correctly associate Go archives with their import path
Data about Go types is inferred from Go archives (either *.x or *.a
files). When a command is rewritten, and type data is needed, this
archive is looked up and parsed by the standard gcimporter library.

We incorrectly used just the package name to find which file is supposed
to go with an import path, which meant that, for example, a lookup for
"crypto/rand" yielded the math/rand.a file. This was also true for
non-stdlib imports.

For stdlib imports, we can simply do full path matching -- "crypto/rand"
should get "crypto/rand.a" and "math/rand" should get "math/rand.a" --
the path must match the import path for the stdlib.

For non-stdlib imports, we need to match their import path (e.g.
"github.com/u-root/gobusybox/src/bazeltest/pkg/log") to the actual
file path (likely "pkg/log/log.a" in this case).

rules_go's GoArchive provider gives us both the import path and the file
path, so they are passed to rewritepkg with --archive
importPath:archiveFilePath.

Signed-off-by: Chris Koch <chrisko@google.com>