bazel: add support for implicit imports
Originally, go_busybox_library was a proxy for an actual go_library by
rewriting just the source files, but reusing the same list of
dependencies and other attributes on a go_library.

With potential implicit imports becoming explicit, we need to explicitly
add dependencies to the list of `deps` passed to `go_library`. That, in
the original structure, is simply impossible.

Instead, uroot_rewrite_ast (now _go_busybox_library) pretends to be an
actual go_library rule by implementing the GoLibrary, GoSource, and
GoArchive providers for rules_go. [See docs
here](https://github.com/bazelbuild/rules_go/blob/master/go/providers.rst).
The go_dep_aspect allows us to collect the list of Targets of all
transitive dependencies of the rule, which we pass whole.

This is indeed overkill: we need only pass the transitive dependencies
that we actually made explicit. However, figuring out from the output of
rewritepkg which dependencies those actually were is not simple, and
bazel doesn't mind being handed too many dependencies. (Plus, since
they're all transitive, they already have to be compiled for this
target. So this is not adding any more build time.)

Signed-off-by: Chris Koch <chrisko@google.com>