bb: partially undo 002c33a to deconflict Init/Main symbols
Before this commit, the generated main.go would create one large
registration of all commands, like this:

    package main

    import (
        ".../cmd/ls"
        ".../cmd/foobar"
    )

    func init() {
        Register("ls", ls.Init, ls.Main)
        Register("foobar", foobar.Init, foobar.Main)
    }

This relied upon the implicit contract that all rewritten commands would
export an Init and Main function. However, if the original command
already had an Init or a Main, we'd generate an additional Init or Main
and compilation would fail.

If we simply mangled the name, we'd have to pass information about what
the Init and Main functions are now called for each command from the
"rewrite commands" to the "generate main.go" step. This is easy in a
vendoring or module world, but hard in the bazel world.

Instead, commands are now rewritten to register themselves (again).

I.e., in the example above, cmd/ls would have:

    package ls

    import (
        "bb.u-root.com/pkg/bb/bbmain"
    )

    func init0() ...
    func main0() ...

    func init() {
      bbmain.Register("ls", init0, main0)
    }

cmd/foobar would look similar, with s/ls/foobar of course.

And the generated main will anonymously every command:

    package main

    import (
        _ ".../cmd/ls"
        _ ".../cmd/foobar"
    )

This way, no information has to be passed around in the build system
from the rewrite-commands to the generate main.go step. However, a new
dependency has to be added to every generated command.

Signed-off-by: Chris Koch <chrisko@google.com>