src/pkg/bb: Use localversion over dependency
if any third-party cmd we pull in to build uses u-root as a dependency,
it will always conflict. When building u-root, u-root is always pulled
in as a local dependency i.e. github.com/u-root/u-root@directory. This
conflicts with every go.mod that is outside of u-root/u-root and depends
on u-root.

Signed-off-by: Christian Walter <christian.walter@9elements.com>