loadFSPackages: do not ignore errors from addPkg
loadFSPackages is currently ignoring errors returned by addPkg.

This results in a very bad outcome: firmware images that are missing commands.
For example, if you corrupt, e.g., date, u-root will still build
a firmware image with date missing.

This has, in turn, resulted in the creation of broken images (found in
the wild for those of us using cpu).

There is now a test in u-root specifically for the case that
this error is ignored (https://github.com/u-root/u-root/pull/2451)
and this fix makes that new test work, where it does not pass
currently.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>