Make Exit an overrideable package variable
In bare metal environments such as Tamago,
os.Exit halts the machine.

Add a variable, Exit, to the bbmain package.
The default value is os.Exit.

Have main() in bbmain run a for loop, which calls
bbmain.Exit with the return value of run().

change run() to return the value, not call os.Exit.

These changes should make no difference in Linux, but in
tinygo and Tamago environments it will be possible
to ensure that the main loop no longer exits.

Signed-off-by: Ron Minnich <rminnich@gmail.com>